--- a/.changeset/short-onions-play.md
+++ b/.changeset/short-onions-play.md
@@ -1,0 +1,5 @@
+---
+"stylelint": patch
+---
+
+Fixed: `selector|value-no-vendor-prefix` performance