--- a/lib/reference/prefixes.js
+++ b/lib/reference/prefixes.js
@@ -1,0 +1,5 @@
+const prefixes = new Set(['-webkit-', '-moz-', '-ms-', '-o-']);
+
+module.exports = {
+	prefixes,
+};