--- a/lib/rules/selector-no-vendor-prefix/index.js
+++ b/lib/rules/selector-no-vendor-prefix/index.js
@@ -1,5 +1,6 @@
 'use strict';
 
+const hasPrefix = require('../../utils/hasPrefix');
 const isAutoprefixable = require('../../utils/isAutoprefixable');
 const isStandardSyntaxRule = require('../../utils/isStandardSyntaxRule');
 const optionsMatches = require('../../utils/optionsMatches');
@@ -41,6 +42,8 @@ const rule = (primary, secondaryOptions, context) => {
 		}
 
 		root.walkRules((ruleNode) => {
+			if (!hasPrefix(ruleNode?.selector)) return;
+
 			if (!isStandardSyntaxRule(ruleNode)) {
 				return;
 			}