--- a/lib/rules/selector-no-vendor-prefix/index.js
+++ b/lib/rules/selector-no-vendor-prefix/index.js
@@ -42,7 +42,7 @@ const rule = (primary, secondaryOptions, context) => {
 		}
 
 		root.walkRules((ruleNode) => {
-			if (!hasPrefix(ruleNode?.selector)) return;
+			if (!hasPrefix(ruleNode.selector)) return;
 
 			if (!isStandardSyntaxRule(ruleNode)) {
 				return;