--- a/lib/rules/value-no-vendor-prefix/index.js
+++ b/lib/rules/value-no-vendor-prefix/index.js
@@ -2,6 +2,7 @@
 
 const valueParser = require('postcss-value-parser');
 
+const hasPrefix = require('../../utils/hasPrefix');
 const isAutoprefixable = require('../../utils/isAutoprefixable');
 const isStandardSyntaxDeclaration = require('../../utils/isStandardSyntaxDeclaration');
 const isStandardSyntaxProperty = require('../../utils/isStandardSyntaxProperty');
@@ -22,18 +23,6 @@ const messages = ruleMessages(ruleName, {
 const meta = {
 	url: 'https://stylelint.io/user-guide/rules/value-no-vendor-prefix',
 	fixable: true,
-};
-
-const valuePrefixes = ['-webkit-', '-moz-', '-ms-', '-o-'];
-
-/**
- * @param {string} value
- * @returns {boolean}
- */
-const hasPrefix = (value) => {
-	const lowerValue = value.toLowerCase();
-
-	return valuePrefixes.some((prefix) => lowerValue.startsWith(prefix));
 };
 
 /** @type {import('stylelint').Rule} */
@@ -59,11 +48,9 @@ const rule = (primary, secondaryOptions, context) => {
 		root.walkDecls((decl) => {
 			const { value } = decl;
 
-			if (
-				!isStandardSyntaxDeclaration(decl) ||
-				!isStandardSyntaxProperty(decl.prop) ||
-				!value.startsWith('-')
-			) {
+			if (!hasPrefix(value)) return;
+
+			if (!isStandardSyntaxDeclaration(decl) || !isStandardSyntaxProperty(decl.prop)) {
 				return;
 			}
 
@@ -74,10 +61,6 @@ const rule = (primary, secondaryOptions, context) => {
 			const parsedValue = valueParser(value);
 
 			parsedValue.walk((node) => {
-				if (!hasPrefix(node.value)) {
-					return;
-				}
-
 				if (!isAutoprefixable.propertyValue(node.value)) {
 					return;
 				}