--- a/lib/utils/hasPrefix.js
+++ b/lib/utils/hasPrefix.js
@@ -2,7 +2,7 @@
 
 const { prefixes } = require('../reference/prefixes');
 
-const HAS_PREFIX_REGEX = new RegExp(`(?:${[...prefixes.values()].join('|')})`, 'i');
+const HAS_PREFIX_REGEX = new RegExp(`(?:${[...prefixes].join('|')})`, 'i');
 
 /**
  * Check if a string contains any prefix