<!-- Each pull request must be associated with an open issue unless it's a documentation fix. If a corresponding issue does not exist, please create one so we can discuss the change first. -->

<!-- Please answer the following. We close pull requests that don't. -->

> Which issue, if any, is this issue related to?

https://github.com/stylelint/stylelint/issues/6869

> Is there anything in the PR that needs further explanation?

selector-no-vendor-prefix

Mean: 246.52507291666666 ms
Deviation: 22.75678614753106 ms

To

Mean: 91.62227779999999 ms
Deviation: 7.8451629248369095 ms


value-no-vendor-prefix

Mean: 102.0994388 ms
Deviation: 11.78154094781758 ms

To

Mean: 91.27013606666667 ms
Deviation: 15.351019516794766 ms